+++ tests/testthat/helper.R
@@ -1,0 +1,4 @@
+
+expect_no_error <- function(object, ...) {
+  expect_error(object, NA, ...)
+}