+++ tests/testthat/test-get_output_data.R
@@ -1,39 +1,50 @@
-with_mock_dir("apsystems", {
-  test_that("get_output_data() works with one APSystems device", {
-    skip_on_cran()
-    expect_error(
-      get_output_data(apsystems_host),
-      NA)
-    apsystem_data <-  get_output_data(apsystems_host)
-    expect_true(is.data.frame(apsystem_data))
-    expect_equal(
-      names(apsystem_data),
-      c("device_id", "inverter", "output_power", "today_energy", "lifetime_energy")
-    )
-    expect_equal(nrow(apsystem_data), 2L)
-  })
 
-  test_that("get_output_data() works with multiple devices from APSystems", {
-    skip_on_cran()
-    expect_error(
-      get_output_data(apsystems_multi),
-      NA)
-    apsystem_data <-  get_output_data(apsystems_multi)
-    expect_true(is.data.frame(apsystem_data))
-    expect_equal(
-      names(apsystem_data),
-      c("device_id", "inverter", "output_power", "today_energy", "lifetime_energy")
-    )
-    expect_equal(nrow(apsystem_data), 4L)
-  })
+local_mocked_bindings(.req_perform_parallel = function(requests, ...) {
+  lapply(requests, httr2::req_perform)
+})
+
+with_mocked_bindings(
+  with_mock_dir("apsystems", {
+    test_that("get_output_data() works with one APSystems device", {
+      skip_on_cran()
+      expect_error(
+        get_output_data(apsystems_host),
+        NA)
+      apsystem_data <-  get_output_data(apsystems_host)
+      expect_true(is.data.frame(apsystem_data))
+      expect_equal(
+        names(apsystem_data),
+        c("device_id", "inverter", "output_power", "today_energy", "lifetime_energy")
+      )
+      expect_equal(nrow(apsystem_data), 2L)
+    })
+
+    test_that("get_output_data() works with multiple devices from APSystems", {
+      skip_on_cran()
+      expect_error(
+        get_output_data(apsystems_multi),
+        NA)
+      apsystem_data <-  get_output_data(apsystems_multi)
+      expect_true(is.data.frame(apsystem_data))
+      expect_equal(
+        names(apsystem_data),
+        c("device_id", "inverter", "output_power", "today_energy", "lifetime_energy")
+      )
+      expect_equal(nrow(apsystem_data), 4L)
+    })
+  }),
+  check_device_ip = function(device_ip) {
+  if (rlang::enexpr(device_ip) %in% c("apsystems_host", "apsystems_multi")) {
+    return
+  }
 })
 
 with_mock_dir("enphase", {
   test_that("get_output_data() works with one Enphase-Energy device", {
     skip_on_cran()
-    expect_error(
-      enphase_data <-  get_output_data(device_ip = "enphase.local", model = "Enphase-Energy"),
-      NA)
+    expect_no_error(
+      enphase_data <-  get_output_data(device_ip = "enphase.local", model = "Enphase-Energy")
+      )
     expect_true(is.data.frame(enphase_data))
     expect_equal(
       names(enphase_data),
@@ -44,9 +55,9 @@ with_mock_dir("enphase", {
 
   test_that("get_output_data() works with one Enphase-Envoy device", {
     skip_on_cran()
-    expect_error(
-      enphase_data <-  get_output_data(device_ip = "enphase.local", model = "Enphase-Envoy"),
-      NA)
+    expect_no_error(
+      enphase_data <-  get_output_data(device_ip = "enphase.local", model = "Enphase-Envoy")
+      )
     expect_true(is.data.frame(enphase_data))
     expect_equal(
       names(enphase_data),
@@ -59,10 +70,10 @@ with_mock_dir("enphase", {
 with_mock_dir("f", {
   test_that("get_output_data() works with one Fronius device", {
     skip_on_cran()
-    expect_error(
-      get_output_data(device_ip = "fronius", model = "Fronius"),
-      NA)
-    fronius_data <-  get_output_data(device_ip = "fronius", model = "Fronius")
+    expect_no_error(
+      get_output_data(device_ip = "f.local", model = "Fronius")
+      )
+    fronius_data <-  get_output_data(device_ip = "f.local", model = "Fronius")
     expect_true(is.data.frame(fronius_data))
     expect_equal(
       names(fronius_data),
@@ -73,10 +84,10 @@ with_mock_dir("f", {
 
   test_that("get_output_data() works with multiple devices from Fronius", {
     skip_on_cran()
-    expect_error(
-      get_output_data(device_ip = c("fronius","fronius2"), model = "Fronius"),
-      NA)
-    fronius_data <-  get_output_data(device_ip = c("fronius","fronius2"), model = "Fronius")
+    expect_no_error(
+      get_output_data(device_ip = c("f.local","g.local"), model = "Fronius")
+      )
+    fronius_data <-  get_output_data(device_ip = c("f.local","g.local"), model = "Fronius")
     expect_true(is.data.frame(fronius_data))
     expect_equal(
       names(fronius_data),
@@ -85,7 +96,6 @@ with_mock_dir("f", {
     expect_equal(nrow(fronius_data), 2L)
   })
 })
-
 
 test_that("get_output_data() raise an explicit message for unsupported model", {
   expect_error(